face-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
insightface>=0.7.3
onnxruntime>=1.16.0
